20260625 업데이트 - KPI/실시간 모니터링 API 추가

+++ fems_fastApi/web/api/energy_usage_pred/schema.py
@@ -19,6 +19,6 @@ class EnergyUsagePredResponse(BaseModel):
     unit: str | None = Field(None, alias="UNIT")
     erroramt: float | None = Field(None, alias="ERRORAMT")
     errorrate: float | None = Field(None, alias="ERRORRATE")
-    accuracy: int | None = Field(None, alias="ACCURACY")
+    accuracy: float | None = Field(None, alias="ACCURACY")
 
     model_config = {"from_attributes": True, "populate_by_name": True}
